Switch the Ruby to browser typescript
This is missing a lot of functionality.
But it gets as far as fetching the login URL for vCloud.

The basic idea is to recreate the Ruby work but in the browser so that I
can see the responses that way.
I have a Rust version as well as this Typescript version. I'm currently
focussing on the Typescript version because I want to assess the
@vcd/bindings package.