Use nginx without a container
This works better on systems that don't support containers so well, such
as Termux on my Chromebook.

It took me a while to figure out how to do relative paths in nginx.
I still don't know how to have it log to stdout in a manner that works
in Termux.