Start writing a mocha test using pollyjs to replay
Currently I'm hitting a couple bugs:
The openapi generated package references btoa, which only exists on the
browser.
I've modified the client as suggested
https://github.com/OpenAPITools/openapi-generator/issues/3049.
As it seems to be working, I'll upstream it.

The pollyjs types don't quite line up with reality, as reality needs to
import *. I'm not sure I understand why this is working this way enough
to say that the bug is in the type definition.